requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.3.0
loguru>=0.7.2
httpx>=0.27.0
sqlalchemy>=2.0.0
redis>=5.0.0
aioredis>=2.0.1
pycryptodome>=3.20.0
apscheduler>=3.10.4
